refactor(opencode): extract shared plugin runtime
