+++ .opencode/plugins/codemem.js
@@ -1,1 +1,1 @@
-export { default } from "../../packages/opencode-plugin/.opencode/plugins/runtime.js";
+export { default } from "../../packages/opencode-plugin/.opencode/plugins/codemem.js";
