--- packages/opencode-plugin/.opencode/plugins/runtime.js
@@ -1,1 +1,0 @@
-export { default } from "./codemem.js";
